# contracts/openapi/foyer-v1.yaml
openapi: 3.1.0
info:
  title: Foyer Server API
  version: 0.4.0
  description: |
    Contract for the self-hosted Foyer Rust service.

    This version covers notes, bookmarks, tasks, contacts, calendar, and device-key
    authentication. Clients communicate only through this contract and
    `contracts/auth/v1/`. They do not receive PostgreSQL credentials, DAV
    credentials, raw iCalendar/vCard payloads, or operation/checkpoint tables.

    Calendar, tasks, and contacts are normalized projections of Radicale. Notes and
    bookmarks are PostgreSQL-canonical. All mutations use client-generated operation
    identifiers. DAV-backed writes also carry the last observed ETag or revision.

    The deleted Cloudflare/Flue HTTP API is not a compatibility target.

    ## Authentication

    Production identity is a manually enrolled ECDSA P-256 device key (ADR 0023).

    - `POST /v1/auth/challenges` with `{ "deviceKeyId": "<thumbprint>" }`.
    - Sign the opaque `signingPayload` bytes with ECDSA P-256/SHA-256.
    - `POST /v1/auth/sessions` with `{ "challengeId", "signature" }` where the
      signature is IEEE P1363 `r||s`, base64url without padding.
    - Send `Authorization: Bearer <accessToken>` on every authenticated route.
    - `GET /v1/auth/jwks` publishes the server ES256 verification key.
    - `GET /v1/sync/credentials` issues a five-minute PowerSync JWT with a
      distinct audience. PowerSync verifies only the public JWKS.

...

...
paths:
...
  /v1/sync/credentials:
    get:
      operationId: getSyncCredentials
      summary: Issue a short-lived PowerSync JWT
      description: |
        Returns a five-minute ES256 JWT and the client-facing PowerSync URL. The
        token uses a distinct audience, is scoped to the current user and device,
        and is not a database credential. PowerSync verifies it with
        `GET /v1/auth/jwks`.
      responses:
        "200":
          description: Credentials the official PowerSync client SDKs can present.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncCredentials"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /v1/calendars:
    get:
      operationId: listCalendars
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createCalendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCalendarRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calendar"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
        "503":
          $ref: "#/components/responses/Unavailable"
...
  /v1/calendars/{calendarId}/rename:
    post:
      operationId: renameCalendar
      parameters:
        - $ref: "#/components/parameters/calendarId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameCalendarRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calendar"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
  /v1/calendars/{calendarId}/delete:
    post:
      operationId: deleteCalendar
      parameters:
        - $ref: "#/components/parameters/calendarId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteCalendarRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calendar"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
        "422":
          $ref: "#/components/responses/SemanticRejection"
  /v1/events:
    get:
      operationId: listEvents
      parameters:
        - name: calendarId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
        "503":
          $ref: "#/components/responses/Unavailable"
...
  /v1/events/{eventId}/update:
    post:
      operationId: updateEvent
      parameters:
        - $ref: "#/components/parameters/eventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
  /v1/events/{eventId}/move:
    post:
      operationId: moveEvent
      parameters:
        - $ref: "#/components/parameters/eventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveEventRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
  /v1/events/{eventId}/delete:
    post:
      operationId: deleteEvent
      parameters:
        - $ref: "#/components/parameters/eventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteCalendarRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
components:
  securitySchemes:
    foyerAccess:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Five-minute ES256 access JWT issued by POST /v1/auth/sessions. Claims
        are documented in contracts/auth/v1/.
    developmentBearer:
      type: http
      scheme: bearer
      description: >
        Development-only static token from FOYER_DEV_TOKEN. Rejected unless the
        process is running with FOYER_SERVER_ENV=development.
  parameters:
...
    calendarId:
      name: calendarId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    eventId:
      name: eventId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Forbidden:
      description: The token is not usable in this process environment.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    NotFound:
      description: The entity does not exist for this user.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Conflict:
      description: Stale expected revision or identifier collision.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Gone:
      description: The entity is tombstoned and cannot be resurrected.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    InvalidRequest:
      description: The request failed input bounds or schema validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    SemanticRejection:
      description: >
        The mutation is well-formed but violates folder emptiness, parent, cycle,
        or limit rules.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Unavailable:
      description: PostgreSQL or the DAV authority is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
...
  schemas:
...
    SyncCredentials:
      type: object
      additionalProperties: false
      required:
        - endpoint
        - token
        - expiresAt
        - userId
      properties:
        endpoint:
          type: string
          format: uri
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
        userId:
          type: string
...
    ErrorBody:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/Error"
    Error:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - unauthenticated
            - forbidden
            - not_found
            - gone
            - stale_revision
            - conflict
            - invalid_request
            - invalid_parent
            - cycle
            - folder_not_empty
            - address_book_not_empty
            - stale_etag
            - limit_exceeded
            - unavailable
            - rate_limited
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    Calendar:
      type: object
      additionalProperties: false
      required:
        - id
        - userId
        - uid
        - href
        - etag
        - displayName
        - description
        - revision
        - createdAt
        - updatedAt
        - deletedAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        uid:
          type: string
        href:
          type: string
        etag:
          type: string
        displayName:
          type: string
        description:
          type: string
        color:
          type:
            - string
            - "null"
        revision:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type:
            - string
            - "null"
          format: date-time
    CalendarList:
      type: object
      additionalProperties: false
      required:
        - calendars
      properties:
        calendars:
          type: array
          items:
            $ref: "#/components/schemas/Calendar"
    Event:
      type: object
      additionalProperties: false
      required:
        - id
        - userId
        - calendarId
        - uid
        - href
        - etag
        - summary
        - description
        - location
        - allDay
        - dtstart
        - revision
        - createdAt
        - updatedAt
        - deletedAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        calendarId:
          type: string
          format: uuid
        uid:
          type: string
        href:
          type: string
        etag:
          type: string
        summary:
          type: string
        description:
          type: string
        location:
          type: string
        allDay:
          type: boolean
        dtstart:
          type: string
        dtend:
          type:
            - string
            - "null"
        tzid:
          type:
            - string
            - "null"
        rrule:
          type:
            - string
            - "null"
        exdates:
          type: string
        revision:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type:
            - string
            - "null"
          format: date-time
    EventList:
      type: object
      additionalProperties: false
      required:
        - events
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
...
    CreateCalendarRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - id
        - displayName
      properties:
        operationId:
          type: string
          format: uuid
        id:
          type: string
          format: uuid
        displayName:
          type: string
        description:
          type: string
        color:
          type: string
    RenameCalendarRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - expectedRevision
        - displayName
      properties:
        operationId:
          type: string
          format: uuid
        expectedRevision:
          type: integer
        expectedEtag:
          type: string
        displayName:
          type: string
    CreateEventRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - id
        - calendarId
        - summary
        - allDay
        - dtstart
      properties:
        operationId:
          type: string
          format: uuid
        id:
          type: string
          format: uuid
        calendarId:
          type: string
          format: uuid
        uid:
          type: string
        summary:
          type: string
        description:
          type: string
        location:
          type: string
        allDay:
          type: boolean
        dtstart:
          type: string
        dtend:
          type: string
        tzid:
          type: string
        rrule:
          type: string
        exdates:
          type: array
          items:
            type: string
    UpdateEventRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - expectedRevision
        - summary
        - allDay
        - dtstart
      properties:
        operationId:
          type: string
          format: uuid
        expectedRevision:
          type: integer
        expectedEtag:
          type: string
        summary:
          type: string
        description:
          type: string
        location:
          type: string
        allDay:
          type: boolean
        dtstart:
          type: string
        dtend:
          type: string
        tzid:
          type: string
        rrule:
          type: string
        exdates:
          type: array
          items:
            type: string
    MoveEventRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - expectedRevision
        - calendarId
      properties:
        operationId:
          type: string
          format: uuid
        expectedRevision:
          type: integer
        expectedEtag:
          type: string
        calendarId:
          type: string
          format: uuid
    DeleteCalendarRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - expectedRevision
      properties:
        operationId:
          type: string
          format: uuid
        expectedRevision:
          type: integer
        expectedEtag:
          type: string
security:
  - foyerAccess: []
  - developmentBearer: []